Refactor position and trade models to use Decimal for financial attributes; add comprehensive tests for positions and trades API

+++ src/position/schemas.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Dict
 
 from pydantic import BaseModel
@@ -5,9 +6,9 @@ from pydantic import BaseModel
 
 class PositionResponse(BaseModel):
     symbol: str
-    quantity: float
-    avg_price: float
-    realized_pnl: float
+    quantity: Decimal
+    avg_price: Decimal
+    realized_pnl: Decimal
 
 
 class AllPositionsResponse(BaseModel):
