+++ src/trades/schemas.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import List
 
 from pydantic import BaseModel
@@ -10,8 +11,8 @@ class TradeCreate(BaseModel):
     id: str
     symbol: str
     side: TradeSide
-    price: float
-    quantity: float
+    price: Decimal
+    quantity: Decimal
     timestamp: datetime
 
 
@@ -19,8 +20,8 @@ class TradeResponse(BaseModel):
     id: str
     symbol: str
     side: TradeSide
-    price: float
-    quantity: float
+    price: Decimal
+    quantity: Decimal
     timestamp: datetime
